backend/requirements.txt
fastapi
uvicorn
google-cloud-vision
python-multipart
azure-ai-documentintelligence==1.0.0